+++ sensu/source/CHANGELOG.md
@@ -6,6 +6,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ## [0.4.0] - 2018-07-13
 - Added handling for client environments set as array of values
 - Added schemas so graphite metric names can be parsed into dimensional metrics
+- Added debug option
 
 ## [0.3.0] - 2018-07-13
 - Better logging